requirements.txt
Django>=5.2,<6.0
djangorestframework==3.16.0
django-cors-headers==4.7.0
gunicorn==23.0.0
mysqlclient==2.2.7
Pillow==11.2.1
asgiref==3.8.1
sqlparse==0.5.3
packaging==25.0
whitenoise==6.9.0
djangorestframework-simplejwt==5.5.0